move get joinedTeamAPI to team.ts
